+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - Consolidate formatter argument extraction.
 - Log warning at 1,000 and info at 100 queued messages if logging appenders back up.
+- SemanticLogger::Appender::Base is now SemanticLogger::Subscriber
 
 ### Fixed
 - `SemanticLogger.on_metric` when using a Symbol was not resolving to the correct namespace.